Rename deprecated file
