+++ src/test-utils/src/index.js
@@ -3,10 +3,19 @@ export {default as gl} from './utils/setup-gl';
 export * from './luma.gl/gpgpu';
 
 // Utilities for update tests (lifecycle tests)
-export {testLayer} from './lifecycle-test-new';
+export {testLayer} from './lifecycle-test';
 
 // Deprecated utilities for update tests (lifecycle tests)
-export * from './lifecycle-test';
+export {
+  testInitializeLayer,
+  testUpdateLayer,
+  testDrawLayer,
+  testLayerUpdates,
+  testSubLayerUpdateTriggers,
+  testCreateLayer,
+  testCreateEmptyLayer,
+  testNullLayer
+} from './lifecycle-test-deprecated';
 
 // Basic utility for rendering multiple scenes (could go into "deck.gl/core")
 export {default as SceneRenderer} from './scene-renderer';
